requirements.txt
uvicorn
fastapi
pydentic
pydantic[email]
pymongo
python-multipart
pyjwt
werkzeug
pillow